# [m3tsz] Fast path decoding for constant time series